+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 2024-09-23
+
+### Added 
+
+- Good path tests for move method in chess
+
+### Changed
+
+- Improved the board print
+
+### Known Issues
+
+- All tests including the board str dont pass, update after adding other pieces
+
 ## [0.1.3] - 2024-09-22
 
 ### Added